Optimize the <ProgressBar /> rendering.
By passing an immutable track instead of a JS object to the <Track />
components, we can use the `PureRenderMixin`, so that the component is
only re-rendered when the values of the state actually have changed.

This way, if we have 100 tracks in the feed, each time that the playing
track progresses, we re-render only one progress bar, instead of 100.