Add an observer to manipulate the Player.
To make sure that we don't have two "sources of truth" (the player and
nuclear's state), the Player is now accessed through an nuclear
observer.

It listens to play and pause requests to ask the player to reflect the
current state (sent by the UI through actions, actions dispatch events,
stores update the state, and finally the observer observes state changes
and make calls to the Player accordingly).