Remove the Playlist Component
This is a pretty bi refactoring, and I'm too lazy to explain it.
Basically, it's better to not try to make components if it means that
the state is dispatched into many components that need to communicate
together because a change of state in one means that the state of some
other component needs to be updated (this was the case with the Playlist
component).

What has been done is try to isolate the pieces of state that really
belong together (playlists containing track ids, each playlist pointing
to a current track, a global current playlist, and a global current
track), and exposing a minimal interface to manipulate this state
(select track, select playlist, next etc.).

Then I duplicated all the Model, Update and View functions, one for the
feed app, one for the radio app. Now I'm trying to create helper
functions to avoid duplicating the code between the two apps (how to
render the navigation, what a Track object should look like etc.).

Let's see how that goes :)