Slightly better async handling
Before this commit, we tried to play youtube videos even though the
player might not be ready.

With this change it's not the case anymore, but the code just keeps
getting messier, it might be time for some refactoring (at least extract
functions and stop relying on variables that should probably not be in
scope).