[radio] remove `Model.playlists` and CustomQueue
The custom queue is buggy and not even displayed in the radio.

Experiment to replace `Model.playlists` by explicit properties
`Model.radio` and `Model.latestTracks`.
It creates a but of duplication, but the code is much easier to follow.
Also, it removes some awkwardness in the view (trying to show the
latest tracks, but getting them from the `playlist` array returns a
`Maybe`, so you need to handle the `Nothing` case even when you *know*
that the playlist is here.